testing article comments using entitlements v6
